Add a translator for metalad_studyminimeta
This translator reuses some jq programs from the studyminimeta
translator (translation schema) written by jsheunis and included in
datalad-catalog, but wraps them in a more explicit python logic.

This makes the translator more similar to other translators included
here, in that empty fields will not be included in the translated
metadata. This also allows us to explicitly in omit publication info
if not present, something that currently causes an error in the
catalog extractor (catalog v0.1.2).