Switch from glob to git ls-files in citation extractors
Previously we used `Path.glob(**/*.ext)` to match files with a given
extension (.ris or .nbib) regardless of their location within the
dataset. This risked consuming an "inordinate amount of time" in large
directory trees, and probably also combed through the `.git`
repository.

The same goal is now achieved with `dataset.repo.call_git_items_()`
used to call `git ls-files *.ext`, and handling of this "search" is
left to git.