Update new cff_translator to match catalog's TranslatorBase
This implements the match method replacing several others, reflecting
changes in TranslatorBase introduced to catalog's main branch.

match() will first check the extractor's uuid (if provided by the
catalog, which may not be the case currently) or name (if uuid is not
passed) then catalog schema version; and finally extractor
version. Version checking is done based on semver, and would fail on
minor but not patch version change - which is probably a bit on the
strict side.

This sets the dependency on catalog's main branch.