Add crossref extractor / translator pair
Added an extractor / translator pair for metadata obtained from
Crossref API. For now we assume that the metadata will be placed in
files with .crossref.json extension, and each file will contain a list
of references upacked from the bigger object returned by the API.

This unpacking of the API response needs to be done by whoever creates
these files, but in the future the extractor could be changed to look
at the object structure and extract the message if needed.

Some more refactoring could also be done now that we have 3 kinds of
citation extractors.